requirements.txt
flask==1.1.1
thor-requests==1.2.5
